requirements.txt
sounddevice
scipy
openai-whisper
numpy
librosa